+++ Core/Inc/utils/iir_filter.h
@@ -12,13 +12,13 @@
 
 typedef struct
 {
-    uint16_t value;
-    uint16_t prev_value;
+    float value;
+    float prev_value;
     float b;   // coefficient
     float sample_time; // ms
 } IIRfilter_t;
 
-void IIR_update(IIRfilter_t *this, uint16_t current_value);
+void IIR_update(IIRfilter_t *this, float current_value);
 void IIR_setCutoffFreq(IIRfilter_t *this, float freq);
 void IIR_reset(IIRfilter_t *this);
